pyproject.toml
[project]
name = "finance-green-agent"
version = "0.1.0"
description = "Financial 10-K Analysis Green Agent for AgentBeats"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
